pyproject.toml
[tool.poetry]
name = "stonks"
version = "0.1.0"
description = ""
authors = ["Vitor <pro.vitoralmeida@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.*"
flask = "^3.0.0"
pydantic = "^2.4.2"
pytest = "^7.4.2"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.5"
flask-wtf = "^1.2.1"
email-validator = "^2.0.0.post2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
